tests, test organizaiton, and improvements for paste
