--- packages/slate/test/commands/at-current-range/insert-fragment/end-block.js
@@ -1,31 +1,0 @@
-/** @jsx h */
-
-import h from '../../../helpers/h'
-
-export default function(editor) {
-  editor.insertFragment(
-    <document>
-      <quote>fragment</quote>
-    </document>
-  )
-}
-
-export const input = (
-  <value>
-    <document>
-      <paragraph>
-        word<cursor />
-      </paragraph>
-    </document>
-  </value>
-)
-
-export const output = (
-  <value>
-    <document>
-      <paragraph>
-        wordfragment<cursor />
-      </paragraph>
-    </document>
-  </value>
-)
