--- packages/slate/test/commands/at-current-range/insert-fragment/fragment-adjacent-texts.js
@@ -1,34 +1,0 @@
-/** @jsx h */
-
-import h from '../../../helpers/h'
-
-export default function(editor) {
-  editor.insertFragment(
-    <document>
-      <paragraph>
-        <text>one</text>
-        <text>two</text>
-      </paragraph>
-    </document>
-  )
-}
-
-export const input = (
-  <value>
-    <document>
-      <paragraph>
-        <cursor />
-      </paragraph>
-    </document>
-  </value>
-)
-
-export const output = (
-  <value>
-    <document>
-      <paragraph>
-        onetwo<cursor />
-      </paragraph>
-    </document>
-  </value>
-)
