--- packages/slate/test/commands/at-current-range/insert-fragment/hanging-selection-single-block.js
@@ -1,42 +1,0 @@
-/** @jsx h */
-
-import h from '../../../helpers/h'
-
-const fragment = (
-  <document>
-    <paragraph>fragment zero</paragraph>
-    <paragraph>fragment one</paragraph>
-    <paragraph>fragment two</paragraph>
-  </document>
-)
-
-export default function(editor) {
-  editor.insertFragment(fragment)
-}
-
-export const input = (
-  <value>
-    <document>
-      <paragraph>zero</paragraph>
-      <paragraph>
-        <anchor />one
-      </paragraph>
-      <quote>
-        <focus />two
-      </quote>
-    </document>
-  </value>
-)
-
-export const output = (
-  <value>
-    <document>
-      <paragraph>zero</paragraph>
-      <quote>fragment zero</quote>
-      <paragraph>fragment one</paragraph>
-      <paragraph>
-        fragment two<cursor />two
-      </paragraph>
-    </document>
-  </value>
-)
