--- packages/slate/test/commands/at-current-range/insert-fragment/merge-deep-nested.js
@@ -1,49 +1,0 @@
-/** @jsx h */
-
-import h from '../../../helpers/h'
-
-export default function(editor) {
-  editor.insertFragment(
-    <document>
-      <code>
-        <quote>
-          <paragraph>2</paragraph>
-          <paragraph>3</paragraph>
-        </quote>
-      </code>
-    </document>
-  )
-}
-
-export const input = (
-  <value>
-    <document>
-      <paragraph>
-        <code>
-          <quote>
-            <paragraph>
-              1<cursor />
-            </paragraph>
-          </quote>
-        </code>
-      </paragraph>
-    </document>
-  </value>
-)
-
-export const output = (
-  <value>
-    <document>
-      <paragraph>
-        <code>
-          <quote>
-            <paragraph>12</paragraph>
-            <paragraph>
-              3<cursor />
-            </paragraph>
-          </quote>
-        </code>
-      </paragraph>
-    </document>
-  </value>
-)
