--- packages/slate/test/commands/at-current-range/insert-fragment/middle-block.js
@@ -1,31 +1,0 @@
-/** @jsx h */
-
-import h from '../../../helpers/h'
-
-export default function(editor) {
-  editor.insertFragment(
-    <document>
-      <quote>fragment</quote>
-    </document>
-  )
-}
-
-export const input = (
-  <value>
-    <document>
-      <paragraph>
-        wo<cursor />rd
-      </paragraph>
-    </document>
-  </value>
-)
-
-export const output = (
-  <value>
-    <document>
-      <paragraph>
-        wofragment<cursor />rd
-      </paragraph>
-    </document>
-  </value>
-)
