--- packages/slate/test/commands/at-current-range/insert-fragment/middle-fragment-adjacent-texts.js
@@ -1,36 +1,0 @@
-/** @jsx h */
-
-import h from '../../../helpers/h'
-
-export default function(editor) {
-  editor.insertFragment(
-    <document>
-      <paragraph>
-        <text>one</text>
-        <text>two</text>
-      </paragraph>
-    </document>
-  )
-}
-
-export const input = (
-  <value>
-    <document>
-      <paragraph>
-        word<cursor />
-      </paragraph>
-      <paragraph>another</paragraph>
-    </document>
-  </value>
-)
-
-export const output = (
-  <value>
-    <document>
-      <paragraph>
-        wordonetwo<cursor />
-      </paragraph>
-      <paragraph>another</paragraph>
-    </document>
-  </value>
-)
