+++ packages/slate/test/commands/at-current-range/insert-fragment/nested-blocks-different-structure/middle-block-nested.js
@@ -1,0 +1,39 @@
+/** @jsx h */
+
+import h from '../../../../helpers/h'
+
+export default function(editor) {
+  editor.insertFragment(
+    <document>
+      <quote>
+        <quote>one</quote>
+        <quote>two</quote>
+      </quote>
+    </document>
+  )
+}
+
+export const input = (
+  <value>
+    <document>
+      <quote>
+        <paragraph>
+          wo<cursor />rd
+        </paragraph>
+      </quote>
+    </document>
+  </value>
+)
+
+export const output = (
+  <value>
+    <document>
+      <quote>
+        <paragraph>woone</paragraph>
+        <quote>
+          two<cursor />rd
+        </quote>
+      </quote>
+    </document>
+  </value>
+)
