+++ packages/slate/test/commands/at-current-range/insert-fragment/nested-blocks-different-structure/middle-block.js
@@ -1,0 +1,39 @@
+/** @jsx h */
+
+import h from '../../../../helpers/h'
+
+export default function(editor) {
+  editor.insertFragment(
+    <document>
+      <quote>
+        <quote>one</quote>
+        <quote>two</quote>
+      </quote>
+    </document>
+  )
+}
+
+export const input = (
+  <value>
+    <document>
+      <paragraph>
+        wo<cursor />rd
+      </paragraph>
+    </document>
+  </value>
+)
+
+export const output = (
+  <value>
+    <document>
+      <paragraph>wo</paragraph>
+      <quote>
+        <quote>one</quote>
+        <quote>
+          two<cursor />
+        </quote>
+      </quote>
+      <paragraph>rd</paragraph>
+    </document>
+  </value>
+)
