--- packages/slate/test/commands/at-current-range/insert-fragment/start-block-multiple-blocks.js
@@ -1,35 +1,0 @@
-/** @jsx h */
-
-import h from '../../../helpers/h'
-
-export default function(editor) {
-  editor.insertFragment(
-    <document>
-      <quote>one</quote>
-      <quote>two</quote>
-      <quote>three</quote>
-    </document>
-  )
-}
-
-export const input = (
-  <value>
-    <document>
-      <paragraph>
-        <cursor />word
-      </paragraph>
-    </document>
-  </value>
-)
-
-export const output = (
-  <value>
-    <document>
-      <paragraph>one</paragraph>
-      <quote>two</quote>
-      <quote>
-        three<cursor />word
-      </quote>
-    </document>
-  </value>
-)
