--- packages/slate/test/commands/at-current-range/insert-fragment/start-block-with-void-no-text.js
@@ -1,38 +1,0 @@
-/** @jsx h */
-
-import h from '../../../helpers/h'
-
-export default function(editor) {
-  editor.insertFragment(
-    <document>
-      <paragraph>
-        <text>one</text>
-        <text>two</text>
-      </paragraph>
-    </document>
-  )
-}
-
-export const input = (
-  <value>
-    <document>
-      <paragraph>
-        <emoji />
-        <cursor />
-      </paragraph>
-    </document>
-  </value>
-)
-
-export const output = (
-  <value>
-    <document>
-      <paragraph>
-        <emoji />
-        <text>
-          onetwo<cursor />
-        </text>
-      </paragraph>
-    </document>
-  </value>
-)
