--- packages/slate/test/commands/at-current-range/insert-fragment/start-inline.js
@@ -1,33 +1,0 @@
-/** @jsx h */
-
-import h from '../../../helpers/h'
-
-export default function(editor) {
-  editor.insertFragment(
-    <document>
-      <quote>fragment</quote>
-    </document>
-  )
-}
-
-export const input = (
-  <value>
-    <document>
-      <paragraph>
-        <cursor />
-        <link>word</link>
-      </paragraph>
-    </document>
-  </value>
-)
-
-export const output = (
-  <value>
-    <document>
-      <paragraph>
-        fragment<cursor />
-        <link>word</link>
-      </paragraph>
-    </document>
-  </value>
-)
