--- packages/slate/test/commands/at-current-range/insert-fragment/start-second-block.js
@@ -1,33 +1,0 @@
-/** @jsx h */
-
-import h from '../../../helpers/h'
-
-export default function(editor) {
-  editor.insertFragment(
-    <document>
-      <quote>fragment</quote>
-    </document>
-  )
-}
-
-export const input = (
-  <value>
-    <document>
-      <paragraph>word</paragraph>
-      <paragraph>
-        <cursor />another
-      </paragraph>
-    </document>
-  </value>
-)
-
-export const output = (
-  <value>
-    <document>
-      <paragraph>word</paragraph>
-      <paragraph>
-        fragment<cursor />another
-      </paragraph>
-    </document>
-  </value>
-)
