+++ packages/slate/test/commands/at-current-range/insert-fragment/words/adjacent-texts-after-void.js
@@ -1,0 +1,38 @@
+/** @jsx h */
+
+import h from '../../../../helpers/h'
+
+export default function(editor) {
+  editor.insertFragment(
+    <document>
+      <paragraph>
+        <text>one</text>
+        <text>two</text>
+      </paragraph>
+    </document>
+  )
+}
+
+export const input = (
+  <value>
+    <document>
+      <paragraph>
+        <emoji />
+        <cursor />
+      </paragraph>
+    </document>
+  </value>
+)
+
+export const output = (
+  <value>
+    <document>
+      <paragraph>
+        <emoji />
+        <text>
+          onetwo<cursor />
+        </text>
+      </paragraph>
+    </document>
+  </value>
+)
