+++ packages/slate/test/commands/at-current-range/insert-fragment/words/after-inline.js
@@ -1,0 +1,33 @@
+/** @jsx h */
+
+import h from '../../../../helpers/h'
+
+export default function(editor) {
+  editor.insertFragment(
+    <document>
+      <quote>fragment</quote>
+    </document>
+  )
+}
+
+export const input = (
+  <value>
+    <document>
+      <paragraph>
+        <link>word</link>
+        <cursor />
+      </paragraph>
+    </document>
+  </value>
+)
+
+export const output = (
+  <value>
+    <document>
+      <paragraph>
+        <link>word</link>
+        fragment<cursor />
+      </paragraph>
+    </document>
+  </value>
+)
