+++ packages/slate/test/commands/at-current-range/insert-fragment/words/before-inline.js
@@ -1,0 +1,33 @@
+/** @jsx h */
+
+import h from '../../../../helpers/h'
+
+export default function(editor) {
+  editor.insertFragment(
+    <document>
+      <quote>fragment</quote>
+    </document>
+  )
+}
+
+export const input = (
+  <value>
+    <document>
+      <paragraph>
+        <cursor />
+        <link>word</link>
+      </paragraph>
+    </document>
+  </value>
+)
+
+export const output = (
+  <value>
+    <document>
+      <paragraph>
+        fragment<cursor />
+        <link>word</link>
+      </paragraph>
+    </document>
+  </value>
+)
