+++ packages/slate/test/commands/at-current-range/insert-fragment/words/start-block.js
@@ -1,0 +1,31 @@
+/** @jsx h */
+
+import h from '../../../../helpers/h'
+
+export default function(editor) {
+  editor.insertFragment(
+    <document>
+      <quote>fragment</quote>
+    </document>
+  )
+}
+
+export const input = (
+  <value>
+    <document>
+      <paragraph>
+        <cursor />word
+      </paragraph>
+    </document>
+  </value>
+)
+
+export const output = (
+  <value>
+    <document>
+      <paragraph>
+        fragment<cursor />word
+      </paragraph>
+    </document>
+  </value>
+)
