+++ packages/slate/test/commands/at-current-range/insert-fragment/words/with-delete-across-blocks.js
@@ -1,0 +1,34 @@
+/** @jsx h */
+
+import h from '../../../../helpers/h'
+
+export default function(editor) {
+  editor.insertFragment(
+    <document>
+      <quote>fragment</quote>
+    </document>
+  )
+}
+
+export const input = (
+  <value>
+    <document>
+      <paragraph>
+        wo<anchor />rd
+      </paragraph>
+      <paragraph>
+        an<focus />other
+      </paragraph>
+    </document>
+  </value>
+)
+
+export const output = (
+  <value>
+    <document>
+      <paragraph>
+        wofragment<cursor />other
+      </paragraph>
+    </document>
+  </value>
+)
